Move htmx controller to views.go
