+++ app/controllers/files.go
@@ -3,9 +3,9 @@ package controllers
 import (
 	"net/http"
 
-	"github.com/kevingil/blog/app/cmd"
+	"github.com/kevingil/blog/app/views"
 )
 
 func Files(w http.ResponseWriter, r *http.Request) {
-	cmd.Hx(w, r, "dashboard", "dashboard-files", data)
+	views.Hx(w, r, "dashboard", "dashboard-files", data)
 }
