+++ app/views/views.go
@@ -1,6 +1,8 @@
 package views
 
 import (
+	"io"
+	"net/http"
 	"path/filepath"
 	"regexp"
 	"text/template"
@@ -16,6 +18,32 @@ import (
 
 // Tmpl is a template.
 var Tmpl *template.Template
+
+// Hx is a function to render a child template wrapped with a specified layout
+func Hx(w http.ResponseWriter, r *http.Request, layout string, tmpl string, data any) {
+	var response bytes.Buffer
+	var child bytes.Buffer
+
+	if err := Tmpl.ExecuteTemplate(&child, tmpl, data); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+
+	if r.Header.Get("HX-Request") == "true" {
+		io.WriteString(w, child.String())
+
+	} else {
+		if err := Tmpl.ExecuteTemplate(&response, layout, child.String()); err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		io.WriteString(w, response.String())
+
+	}
+
+}
 
 func until(n int) []struct{} {
 	return make([]struct{}, n)
